Made change to small error in users.py

+++ server/models/users.py
@@ -4,7 +4,7 @@ from .base_case import base_case as Base
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer(15), primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     name = Column(String(30), nullable=False)
     email = Column(String(30), unique=True, nullable=False)
     password = Column(String(20), nullable=False)
